Found & corrected the error.
The "showPage()" function call was not working because the function itself was deleted during the clean-up during class session when we were removing what was needed and what not. This was deleted too together with the "hidePage()" function because, I thought that we will rewrite that too, then I forgot about it.